clean up language button
